generate correlation features for replogle

+++ src/data/perturb_dataset.py
@@ -12,6 +12,7 @@ from torch.utils.data import Dataset
 from sklearn.model_selection import train_test_split
 from tqdm import tqdm
 from hydra.errors import HydraException
+from scipy.stats import pearsonr
 
 from src.utils.spectra import get_splits
 
@@ -39,24 +40,42 @@ class PerturbData(Dataset):
             repl = len(sps)
             self.spectral_parameter = "{:.2f}_{}".format(self.spectral_parameter, repl)
 
-        if not os.path.exists(f"{self.data_path}/input_features/train_data_{spectral_parameter}.pkl.gz"):
-            pp_data = self.preprocess_and_featurise_norman(adata)
-            self.X_train, self.train_target, self.X_val, self.val_target, self.X_test, self.test_target = pp_data
-        else:
-            with gzip.open(f"{self.data_path}/input_features/train_data_{spectral_parameter}.pkl.gz", "rb") as f:
-                self.X_train, self.train_target = pkl.load(f)
-            with gzip.open(f"{self.data_path}/input_features/val_data_{spectral_parameter}.pkl.gz", "rb") as f:
-                self.X_val, self.val_target = pkl.load(f)
-            with gzip.open(f"{self.data_path}/input_features/test_data_{spectral_parameter}.pkl.gz", "rb") as f:
-                self.X_test, self.test_target = pkl.load(f)
+        if self.data_name == "norman":
+            if not os.path.exists(f"{self.data_path}/input_features/train_data_{spectral_parameter}.pkl.gz"):
+                pp_data = self.preprocess_and_featurise_norman(adata)
+                self.X_train, self.train_target, self.X_val, self.val_target, self.X_test, self.test_target = pp_data
+            else:
+                with gzip.open(f"{self.data_path}/input_features/train_data_{spectral_parameter}.pkl.gz", "rb") as f:
+                    self.X_train, self.train_target = pkl.load(f)
+                with gzip.open(f"{self.data_path}/input_features/val_data_{spectral_parameter}.pkl.gz", "rb") as f:
+                    self.X_val, self.val_target = pkl.load(f)
+                with gzip.open(f"{self.data_path}/input_features/test_data_{spectral_parameter}.pkl.gz", "rb") as f:
+                    self.X_test, self.test_target = pkl.load(f)
 
         if self.data_name == "replogle_rpe1":
-            ctrl_adata, pert_adata, train, test, pert_list = self.preprocess_replogle(adata)
-            self.featurise_replogle(pert_adata, pert_list, ctrl_adata, train, test)
-
+            if not os.path.exists(f"{self.data_path}/input_features/train_data_{spectral_parameter}.pkl.gz"):
+                ctrl_adata, pert_adata, train, test, pert_list = self.preprocess_replogle(adata)
+                pp_data = self.featurise_replogle(pert_adata, pert_list, ctrl_adata, train, test)
+                self.X_train, self.train_target, self.X_val, self.val_target, self.X_test, self.test_target = pp_data
+            else:
+                with gzip.open(f"{self.data_path}/input_features/train_data_{spectral_parameter}.pkl.gz", "rb") as f:
+                    self.X_train, self.train_target = pkl.load(f)
+                with gzip.open(f"{self.data_path}/input_features/val_data_{spectral_parameter}.pkl.gz", "rb") as f:
+                    self.X_val, self.val_target = pkl.load(f)
+                with gzip.open(f"{self.data_path}/input_features/test_data_{spectral_parameter}.pkl.gz", "rb") as f:
+                    self.X_test, self.test_target = pkl.load(f)
         if self.data_name == "replogle_k562":
-            ctrl_adata, pert_adata, train, test, pert_list = self.preprocess_replogle(adata)
-            self.featurise_replogle(pert_adata, pert_list, ctrl_adata, train, test)
+            if not os.path.exists(f"{self.data_path}/input_features/train_data_{spectral_parameter}.pkl.gz"):
+                ctrl_adata, pert_adata, train, test, pert_list = self.preprocess_replogle(adata)
+                pp_data = self.featurise_replogle(pert_adata, pert_list, ctrl_adata, train, test)
+                self.X_train, self.train_target, self.X_val, self.val_target, self.X_test, self.test_target = pp_data
+            else:
+                with gzip.open(f"{self.data_path}/input_features/train_data_{spectral_parameter}.pkl.gz", "rb") as f:
+                    self.X_train, self.train_target = pkl.load(f)
+                with gzip.open(f"{self.data_path}/input_features/val_data_{spectral_parameter}.pkl.gz", "rb") as f:
+                    self.X_val, self.val_target = pkl.load(f)
+                with gzip.open(f"{self.data_path}/input_features/test_data_{spectral_parameter}.pkl.gz", "rb") as f:
+                    self.X_test, self.test_target = pkl.load(f)
 
     def preprocess_and_featurise_norman(self, adata):
         nonzero_genes = (adata.X.sum(axis=0) > 5).A1
@@ -158,10 +177,20 @@ class PerturbData(Dataset):
             basal_ctrl_adata = sc.read_h5ad(f"{self.data_path}/basal_ctrl_{self.data_name}_pp_filtered.h5ad")
             pert_adata = sc.read_h5ad(f"{self.data_path}/{self.data_name}_pp_pert_filtered.h5ad")
 
+        # these just need to be the same between datasets, irrespective of order
         control_genes = basal_ctrl_adata.var.index.to_list()
         pert_genes = pert_adata.var.index.to_list()
+
+        # these need to be paired between datasets, and in the same order
+        pert_cell_conditions = pert_adata.obs['condition'].to_list()
+        ctrl_cell_conditions = basal_ctrl_adata.obs['condition'].to_list()
+
         assert control_genes == pert_genes, ("Watch out! Genes in control and perturbation datasets are not the"
                                              " same, or are not indexed the same.")
+
+        assert ctrl_cell_conditions == pert_cell_conditions, ("Watch out! Cell conditions in control and perturbation "
+                                                              "datasets are not the or same, or are not indexed the "
+                                                              "same!")
 
         train_perts = [pert_list[i] for i in train]
         test_perts = [pert_list[i] for i in test]
@@ -241,15 +270,6 @@ class PerturbData(Dataset):
         return X_train, train_target, X_val, val_target, X_test, test_target
 
     def preprocess_replogle(self, adata):
-        if not os.path.exists(f"{self.data_path}/{self.data_name}_filtered.h5ad"):
-            adata.write(f"{self.data_path}/{self.data_name}_raw_counts.h5ad", compression='gzip')
-            adata.layers["counts"] = adata.X.copy()
-            sc.pp.normalize_total(adata)
-            sc.pp.log1p(adata)
-            adata.write(f"{self.data_path}/{self.data_name}_filtered.h5ad", compression='gzip')
-        else:
-            adata = sc.read_h5ad(f"{self.data_path}/{self.data_name}_filtered.h5ad")
-
         adata.obs['condition'] = adata.obs['perturbation'].replace('control', 'ctrl')
 
         genes = adata.var.index.to_list()
@@ -257,39 +277,49 @@ class PerturbData(Dataset):
 
         # we remove the cells with perts that are not in the genes because we need gene expression values
         # to generate an in-silico perturbation embedding
-        sg_pert_adata = adata[adata.obs['condition'].isin(genes_and_ctrl), :]
+        adata = adata[adata.obs['condition'].isin(genes_and_ctrl), :]
+        unique_perts = list(set(adata.obs['condition'].to_list()))
+        unique_perts.remove('ctrl')
 
-        ctrl_adata = sg_pert_adata[sg_pert_adata.obs['condition'] == 'ctrl', :]
-        pert_adata = sg_pert_adata[sg_pert_adata.obs['condition'] != 'ctrl', :]
-        all_perts = list(set(pert_adata.obs['condition'].to_list()))
+        sc.pp.normalize_total(adata)
+        sc.pp.log1p(adata)
+        sc.pp.highly_variable_genes(adata, n_top_genes=2000)
+        highly_variable_genes = list(adata.var_names[adata.var['highly_variable']])
+        missing_perts = list(set(unique_perts) - set(highly_variable_genes))
+        combined_genes = list(set(highly_variable_genes) | set(missing_perts))
+        adata = adata[:, combined_genes]
+
+        ctrl_adata = adata[adata.obs['condition'] == 'ctrl', :]
+        pert_adata = adata[adata.obs['condition'] != 'ctrl', :]
 
         num_cells = ctrl_adata.shape[0]
-        num_perts = len(all_perts)
+        num_perts = len(unique_perts)
         mask = np.zeros((num_cells, num_perts), dtype=bool)
 
-        for i, pert in tqdm(enumerate(all_perts), total=len(all_perts)):
-            pert_idx = genes.index(pert)
-            non_zero_indices = ctrl_adata[:, pert_idx].X.sum(axis=1).nonzero()[0]
-            num_non_zeroes = len(non_zero_indices)
+        if not os.path.exists(f"{self.data_path}/{self.data_name}_mask_df.pkl"):
+            for i, pert in tqdm(enumerate(unique_perts), total=len(unique_perts)):
+                pert_idx = genes.index(pert)
+                non_zero_indices = ctrl_adata[:, pert_idx].X.sum(axis=1).nonzero()[0]
+                num_non_zeroes = len(non_zero_indices)
 
-            if len(non_zero_indices) < 500:
-                sample_num = num_non_zeroes
-            else:
-                sample_num = 500
+                if len(non_zero_indices) < 500:
+                    sample_num = num_non_zeroes
+                else:
+                    sample_num = 500
 
-            sampled_indices = np.random.choice(non_zero_indices, sample_num, replace=False)
+                sampled_indices = np.random.choice(non_zero_indices, sample_num, replace=False)
 
-            mask[sampled_indices, i] = True
+                mask[sampled_indices, i] = True
 
-        mask_df = pd.DataFrame(mask, columns=all_perts)
+            mask_df = pd.DataFrame(mask, columns=unique_perts)
 
-        mask_df.to_pickle(f"{self.data_path}/{self.data_name}_mask_df.pkl")
+            mask_df.to_pickle(f"{self.data_path}/{self.data_name}_mask_df.pkl")
 
         if not os.path.exists(f"{self.data_path}/all_perts.pkl"):
             with open(f"{self.data_path}/all_perts.pkl", "wb") as f:
-                pkl.dump(all_perts, f)
+                pkl.dump(unique_perts, f)
 
-        train, test, pert_list = get_splits.spectra(sg_pert_adata,
+        train, test, pert_list = get_splits.spectra(adata,
                                                     self.data_path,
                                                     self.spectra_params,
                                                     self.spectral_parameter
@@ -298,8 +328,6 @@ class PerturbData(Dataset):
 
     def featurise_replogle(self, pert_adata, pert_list, ctrl_adata, train, test):
         print(f"{self.data_name} dataset has {len(pert_list)} perturbations in common with the genes in the dataset.")
-
-        all_perts = pert_adata.obs['condition'].to_list()
 
         if not os.path.exists(f"{self.data_path}/basal_ctrl_{self.data_name}_filtered.h5ad"):
             ctrl_X = ctrl_adata.X.toarray()
@@ -310,6 +338,7 @@ class PerturbData(Dataset):
                 subset = ctrl_X[np.random.choice(ctrl_X.shape[0], subset_size), :]
                 basal_ctrl_X[cell, :] = subset.mean(axis=0)
 
+            # we add pert_adata to obs because we want to pair control expression to perturbed cells
             basal_ctrl_adata = anndata.AnnData(X=basal_ctrl_X, obs=pert_adata.obs, var=ctrl_adata.var)
 
             # noinspection PyTypeChecker
@@ -317,34 +346,69 @@ class PerturbData(Dataset):
         else:
             basal_ctrl_adata = sc.read_h5ad(f"{self.data_path}/basal_ctrl_{self.data_name}_filtered.h5ad")
 
-        control_genes = basal_ctrl_adata.var.index.to_list()
-        pert_genes = pert_adata.var.index.to_list()
-        assert control_genes == pert_genes, ("Watch out! Genes in control and perturbation datasets are not the"
-                                             " same, or are not indexed the same.")
+        # these just need to be the same between datasets, irrespective of order
+        control_genes = sorted(basal_ctrl_adata.var_names.to_list())
+        pert_genes = sorted(pert_adata.var_names.to_list())
+
+        # these need to be paired between datasets, and in the same order
+        pert_cell_conditions = pert_adata.obs['condition'].to_list()
+        ctrl_cell_conditions = basal_ctrl_adata.obs['condition'].to_list()
+
+        assert control_genes == pert_genes, "Watch out! Genes in control and perturbation datasets do not match!"
+
+        assert ctrl_cell_conditions == pert_cell_conditions, ("Watch out! Cell conditions in control and perturbation "
+                                                              "datasets are not the or same, or are not indexed the "
+                                                              "same!")
 
         train_perts = [pert_list[i] for i in train]
         test_perts = [pert_list[i] for i in test]
 
-        sc.pp.highly_variable_genes(pert_adata, n_top_genes=5000)
-        highly_variable_genes = pert_adata.var_names[pert_adata.var['highly_variable']]
-        hv_pert_adata = pert_adata[:, highly_variable_genes]
-
-        train_target = hv_pert_adata[hv_pert_adata.obs['condition'].isin(train_perts), :]
-        test_target = hv_pert_adata[hv_pert_adata.obs['condition'].isin(test_perts), :]
+        train_target = pert_adata[pert_adata.obs['condition'].isin(train_perts), :]
+        test_target = pert_adata[pert_adata.obs['condition'].isin(test_perts), :]
 
         all_perts_train = train_target.obs['condition'].values
         all_perts_test = test_target.obs['condition'].values
 
-        perts_idx = {}
-        for pert in all_perts:
-            perts_idx[pert] = pert_genes.index(pert)
+        if not os.path.exists(f"{self.data_path}/pert_corrs.pkl"):
+            all_gene_expression = basal_ctrl_adata.X
+
+            results = []
+            for pert in tqdm(pert_list, total=len(pert_list)):
+                pert, correlations = self.compute_correlations(pert, basal_ctrl_adata, all_gene_expression)
+                results.append((pert, correlations))
+
+            pert_corrs = {pert: corr for pert, corr in results}
+
+            with open(f"{self.data_path}/pert_corrs.pkl", "wb") as f:
+                pkl.dump(pert_corrs, f)
+        else:
+            with open(f"{self.data_path}/pert_corrs.pkl", "rb") as f:
+                pert_corrs = pkl.load(f)
 
         num_ctrl_cells = basal_ctrl_adata.shape[0]
         num_train_cells = train_target.shape[0]
         num_test_cells = test_target.shape[0]
         num_genes = basal_ctrl_adata.shape[1]
 
+        pert_corr_train = np.zeros((num_train_cells, num_genes))
+        for i, pert in tqdm(enumerate(all_perts_train), total=len(all_perts_train)):
+            pert_corr_train[i, :] = pert_corrs[pert]
+
+        pert_corr_test = np.zeros((num_test_cells, num_genes))
+        for i, pert in tqdm(enumerate(all_perts_test), total=len(all_perts_test)):
+            pert_corr_test[i, :] = pert_corrs[pert]
+
+
         pass  # continue here
+
+    @staticmethod
+    def compute_correlations(pert, basal_ctrl_adata, all_gene_expression):
+        pert_idx = basal_ctrl_adata.var_names.get_loc(pert)
+        basal_expr_pert = basal_ctrl_adata.X[:, pert_idx].flatten()
+        correlations = np.array(
+            [pearsonr(basal_expr_pert, all_gene_expression[:, i])[0] for i in range(all_gene_expression.shape[1])])
+        correlations[np.isnan(correlations)] = 0
+        return pert, correlations
 
     def __getitem__(self, index):
         if self.stage == "train":
